Implemented Speech to Text
Added python file that sends string via UART to STM32.

Added functions in STM code that receives the string from python code one char at a time. Currently it just echos what is sent to it, but in the future the functions can be used to spell words in sign